Refactor GCP module to use a request factory
Let's mirror the Google PIpelines implementation to use a RequestFactory and a RequestHandler.
