+++ supportedBackends/google/pipelines/batch/src/main/scala/cromwell/backend/google/pipelines/batch/GcpBatchBackendSingletonActor.scala
@@ -1,21 +1,23 @@
 package cromwell.backend.google.pipelines.batch
 
 import akka.actor.{Actor, ActorLogging, Props}
-import com.google.cloud.batch.v1.{BatchServiceClient, GetJobRequest, JobName}
+import com.google.cloud.batch.v1.JobName
 import cromwell.backend.BackendSingletonActorAbortWorkflow
+import cromwell.backend.google.pipelines.batch.api.GcpBatchRequestFactory
 import cromwell.core.Dispatcher.BackendDispatcher
 
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.{Failure, Success}
 
 object GcpBatchBackendSingletonActor {
-  def props(name: String) = Props(new GcpBatchBackendSingletonActor(name)).withDispatcher(BackendDispatcher)
+  def props(requestFactory: GcpBatchRequestFactory)(implicit requestHandler: GcpBatchApiRequestHandler) = Props(new GcpBatchBackendSingletonActor(requestFactory))
+    .withDispatcher(BackendDispatcher)
 
   // This is the only type of messages that can be processed by this actor from this actor
   sealed trait Action extends Product with Serializable
   object Action {
     final case class SubmitJob(request: GcpBatchRequest) extends Action
-    final case class QueryJobStatus(jobName: JobName) extends Action
+    final case class QueryJob(jobName: JobName) extends Action
   }
 
   // This is the only type of messages produced from this actor while reacting to received messages
@@ -29,32 +31,20 @@ object GcpBatchBackendSingletonActor {
 }
 
 // TODO: Alex - serviceRegistryActor required to do instrumentation
-final class GcpBatchBackendSingletonActor (name: String) extends Actor with ActorLogging {
+final class GcpBatchBackendSingletonActor(requestFactory: GcpBatchRequestFactory)(implicit requestHandler: GcpBatchApiRequestHandler) extends Actor with ActorLogging {
 
   import GcpBatchBackendSingletonActor._
 
   implicit val ec: ExecutionContext = context.dispatcher
 
-  private def queryStatus(name: JobName) = {
-    val request = GetJobRequest.newBuilder.setName(name.toString).build
-    // TODO: Alex - Consider creating this client only once in the app lifecycle
-    val batchServiceClient = BatchServiceClient.create
-    try {
-      batchServiceClient.getJob(request)
-    } finally  {
-      batchServiceClient.close()
-    }
-  }
-
   // TODO: Alex - Find the usefulness for the actor because we are not holding any mutable state in it
   override def receive: Receive = {
     case Action.SubmitJob(request) =>
-      // TODO: Alex - Consider not hardcoding machineType
-      val job = GcpBatchJob(request,"n1-standard-4")
       val replyTo = sender()
       log.info(s"Submitting job (${request.jobName}) to GCP, workflowId = ${request.workflowId}")
       Future {
-        job.submitJob()
+        // TODO: Alex - Consider not hardcoding machineType
+        requestHandler.submit(requestFactory.submitRequest("n1-standard-4", request))
       }.onComplete {
         case Failure(exception) =>
           log.error(exception, s"Failed to submit job (${request.jobName}) to GCP, workflowId = ${request.workflowId}")
@@ -65,11 +55,11 @@ final class GcpBatchBackendSingletonActor (name: String) extends Actor with Acto
           replyTo ! Event.JobSubmitted(job)
       }
 
-    case Action.QueryJobStatus(jobName) =>
+    case Action.QueryJob(jobName) =>
       val replyTo = sender()
 
       Future {
-        queryStatus(jobName)
+        requestHandler.query(requestFactory.queryRequest(jobName))
       }.onComplete {
         case Success(job) =>
           log.info(s"Job status ($jobName) retrieved from GCP, state = ${job.getStatus.getState}")
@@ -84,9 +74,5 @@ final class GcpBatchBackendSingletonActor (name: String) extends Actor with Acto
     case BackendSingletonActorAbortWorkflow(workflowId) =>
       // TODO: Alex - how do we handle this?
       log.info(s"Cromwell requested to abort workflow $workflowId")
-
-    case other =>
-      log.error("Unknown message to GCP Batch Singleton Actor: {}. Dropping it.", other)
-
   }
 }
