+++ supportedBackends/google/pipelines/batch/src/main/scala/cromwell/backend/google/pipelines/batch/GcpBatchRequest.scala
@@ -1,6 +1,6 @@
 package cromwell.backend.google.pipelines.batch
 
-import cromwell.backend.google.pipelines.batch.GcpBatchRequestFactory.CreatePipelineParameters
+import cromwell.backend.google.pipelines.batch.api.GcpBatchRequestFactory.CreatePipelineParameters
 import cromwell.core.WorkflowId
 
 case class GcpBatchRequest(workflowId: WorkflowId,
